requirements.txt
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
tomli==2.0.1
tqdm==4.64.0